+++ mpp/conf.py
@@ -1,16 +1,12 @@
 from PyQt5 import QtWidgets
 from configparser import ConfigParser
 from .ui import Ui_config
-from .utils import getAppDataDir, isLinux
-from os import path, execl
-from qt_material import apply_stylesheet, list_themes
-from PyQt5.QtGui import QIcon
-import sys
+from .utils import getAppDataDir
+from os import path
+from PyQt5.QtCore import QCoreApplication
 
 config_dir = getAppDataDir()
-default_theme = 'system'
-if not isLinux():
-    default_theme = 'default'
+_translate = QCoreApplication.translate
 
 
 def getConf():
@@ -24,6 +20,8 @@ def getConf():
         for name, value in parser.items('mpp'):
             if name == 'update_on_init':
                 value = parser.getboolean('mpp', 'update_on_init')
+            if name == 'rename_download':
+                value = parser.getboolean('mpp', 'rename_download')
             config[name] = value
 
     else:
@@ -34,12 +32,12 @@ def getConf():
         parser.add_section('mpp')
         config = {
             'update_on_init': 1,
+            'rename_download': 0,
             'download_folder': download_dir,
-            'theme': default_theme
         }
         parser.set('mpp', 'update_on_init', '1')
+        parser.set('mpp', 'rename_download', '0')
         parser.set('mpp', 'download_folder', download_dir)
-        parser.set('mpp', 'theme', default_theme)
         with open(config_dir + 'mpp.ini', 'w') as configfile:
             parser.write(configfile)
 
@@ -57,58 +55,46 @@ class configDialog(QtWidgets.QDialog):
         self.ui.setupUi(self)
 
         self.ui.buttonBox.accepted.connect(self.saveConf)
-        self.ui.buttonBox.rejected.connect(self.resetConf)
+        self.ui.selDownFolder.clicked.connect(self.selDownFolder)
 
         self.conf = getConf()
-        self.current_theme = self.conf['theme']
 
         if self.conf['update_on_init']:
             self.ui.cbUpdateInit.setChecked(True)
         else:
             self.ui.cbUpdateInit.setChecked(False)
 
-        self.ui.downFolderEdit.setText(self.conf['download_folder'])
-
-        if isLinux():
-            self.ui.themeLabel.setHidden(True)
-            self.ui.themeSelector.setHidden(True)
+        if 'rename_download' in self.conf and self.conf['rename_download']:
+            self.ui.cbRenameDown.setChecked(True)
         else:
-            themes = ['default'] + list_themes()
-            for theme in themes:
-                self.ui.themeSelector.addItem(theme)
+            self.ui.cbRenameDown.setChecked(False)
 
-            self.ui.themeSelector.setCurrentText(self.conf['theme'])
-            self.ui.themeSelector.currentIndexChanged.connect(self.changeTheme)
+        self.ui.downFolderEdit.setText(self.conf['download_folder'])
 
     def saveConf(self):
         update_on_init = '0'
         if self.ui.cbUpdateInit.isChecked():
             update_on_init = '1'
 
+        if self.ui.cbRenameDown.isChecked():
+            rename_download = '1'
+
         parser = ConfigParser()
         parser.add_section('mpp')
         parser.set('mpp', 'update_on_init', update_on_init)
+        parser.set('mpp', 'rename_download', rename_download)
         parser.set('mpp', 'download_folder', self.ui.downFolderEdit.text())
-        parser.set('mpp', 'theme', self.ui.themeSelector.currentText())
         with open(config_dir + 'mpp.ini', 'w') as configfile:
             parser.write(configfile)
 
-        theme = self.ui.themeSelector.currentText()
-        if theme == 'system' and self.current_theme != 'system':
-            execl(sys.executable, sys.executable, *sys.argv)
-
         return getConf()
 
-    def changeTheme(self, i):
-        theme = self.ui.themeSelector.currentText()
-        apply_stylesheet(self.parent, theme=theme)
-        if theme.find('light') != -1:
-            QIcon.setThemeName('mpp')
-        else:
-            QIcon.setThemeName('mpp-dark')
-
-        self.parent.repaint()
-
-    def resetConf(self):
-        theme = self.conf['theme']
-        apply_stylesheet(self.parent, theme=theme)
+    def selDownFolder(self):
+        folder = str(
+            QtWidgets.QFileDialog.getExistingDirectory(
+                self,
+                _translate('MainWindow', 'Select downloads folder')
+            )
+        )
+        if folder:
+            self.ui.downFolderEdit.setText(folder)