+++ mpp/locales/es.ts
@@ -4,99 +4,104 @@
 <context>
     <name>MainWindow</name>
     <message>
-        <location filename="../ui/Ui_gui.py" line="285"/>
+        <location filename="../ui/Ui_gui.py" line="290"/>
         <source>Podcast</source>
         <translation>Podcast</translation>
     </message>
     <message>
-        <location filename="../ui/Ui_gui.py" line="286"/>
+        <location filename="../ui/Ui_gui.py" line="291"/>
         <source>Episode Title</source>
         <translation>Titulo del episodio</translation>
     </message>
     <message>
-        <location filename="../ui/Ui_gui.py" line="288"/>
+        <location filename="../ui/Ui_gui.py" line="293"/>
         <source>...</source>
         <translation></translation>
     </message>
     <message>
-        <location filename="../ui/Ui_gui.py" line="289"/>
+        <location filename="../ui/Ui_gui.py" line="294"/>
         <source>Podcasts</source>
         <translation>Podcasts</translation>
     </message>
     <message>
-        <location filename="../ui/Ui_gui.py" line="290"/>
+        <location filename="../ui/Ui_gui.py" line="295"/>
         <source>Web</source>
         <translation>Web</translation>
     </message>
     <message>
-        <location filename="../ui/Ui_gui.py" line="291"/>
+        <location filename="../ui/Ui_gui.py" line="296"/>
         <source>Description</source>
         <translation>Descripcion</translation>
     </message>
     <message>
-        <location filename="../ui/Ui_gui.py" line="292"/>
+        <location filename="../ui/Ui_gui.py" line="297"/>
         <source>Home</source>
         <translation>Inicio</translation>
     </message>
     <message>
-        <location filename="../ui/Ui_gui.py" line="293"/>
+        <location filename="../ui/Ui_gui.py" line="298"/>
         <source>Queue</source>
         <translation>Lista de espera</translation>
     </message>
     <message>
-        <location filename="../ui/Ui_gui.py" line="294"/>
+        <location filename="../ui/Ui_gui.py" line="299"/>
         <source>Downloads</source>
         <translation>Descargas</translation>
     </message>
     <message>
-        <location filename="../__init__.py" line="48"/>
+        <location filename="../__init__.py" line="51"/>
         <source>Add podcast</source>
         <translation>Añadir podcast</translation>
     </message>
     <message>
-        <location filename="../__init__.py" line="52"/>
+        <location filename="../__init__.py" line="58"/>
         <source>Update</source>
         <translation>Actualizar</translation>
     </message>
     <message>
-        <location filename="../__init__.py" line="56"/>
+        <location filename="../__init__.py" line="65"/>
         <source>Configure</source>
         <translation>Configurar</translation>
     </message>
     <message>
-        <location filename="../__init__.py" line="67"/>
+        <location filename="../__init__.py" line="79"/>
         <source>Episode</source>
         <translation>Episodio</translation>
     </message>
     <message>
-        <location filename="../__init__.py" line="67"/>
+        <location filename="../__init__.py" line="79"/>
         <source>Published</source>
         <translation>Publicado</translation>
     </message>
     <message>
-        <location filename="../__init__.py" line="67"/>
+        <location filename="../__init__.py" line="79"/>
         <source>Duration</source>
         <translation>Duración</translation>
     </message>
     <message>
-        <location filename="../__init__.py" line="272"/>
+        <location filename="../__init__.py" line="313"/>
         <source>Add to queue</source>
         <translation>Añadir a la cola</translation>
     </message>
     <message>
-        <location filename="../__init__.py" line="287"/>
+        <location filename="../__init__.py" line="332"/>
         <source>Searching new episodes....</source>
         <translation>Buscando nuevos episodios....</translation>
     </message>
     <message>
-        <location filename="../__init__.py" line="303"/>
+        <location filename="../__init__.py" line="352"/>
         <source>Remove from queue</source>
         <translation>Elimilar de la cola</translation>
     </message>
     <message>
-        <location filename="../__init__.py" line="329"/>
+        <location filename="../__init__.py" line="382"/>
         <source>Unsubscribe</source>
         <translation>Desuscribir</translation>
+    </message>
+    <message>
+        <location filename="../conf.py" line="80"/>
+        <source>Select downloads folder</source>
+        <translation>Seleccionar carpeta de descargas</translation>
     </message>
 </context>
 <context>
@@ -130,29 +135,29 @@
 <context>
     <name>configDialog</name>
     <message>
-        <location filename="../ui/Ui_config.py" line="61"/>
+        <location filename="../ui/Ui_config.py" line="58"/>
         <source>Configure</source>
         <translation>Configuración</translation>
     </message>
     <message>
-        <location filename="../ui/Ui_config.py" line="62"/>
+        <location filename="../ui/Ui_config.py" line="60"/>
         <source>Update at init.</source>
         <translation>Actualizar al iniciar.</translation>
     </message>
     <message>
-        <location filename="../ui/Ui_config.py" line="63"/>
+        <location filename="../ui/Ui_config.py" line="61"/>
         <source>Download folder:</source>
         <translation>Carpeta de descargas:</translation>
     </message>
     <message>
-        <location filename="../ui/Ui_config.py" line="64"/>
+        <location filename="../ui/Ui_config.py" line="59"/>
         <source>Select</source>
         <translation>Seleccionar</translation>
     </message>
     <message>
-        <location filename="../ui/Ui_config.py" line="65"/>
-        <source>Theme:</source>
-        <translation>Tema:</translation>
+        <location filename="../ui/Ui_config.py" line="62"/>
+        <source>Rename episodes after download</source>
+        <translation>Renombrar episodios despues de descargar</translation>
     </message>
 </context>
 </TS>